feat: Refactor admin and FAQ routes
